Refactor Graphics.Vector out of the apps
The Graphics module now contains all functionality that was previously
used directly from Graphics.Vector. This allows for easier access to
both the canvas and the context, the app no longer needs to know which
to use. Also all data types (Except for the transform) are now expressed
in Vector2.