Refactor app creation
Move the call to the app constructor to main, this allows the apps to
only know about the app specification type. This type is extracted to a
Toolkit module that provides all tools needed to define an application.
The old CanvasApp type is no longer needed as it is only used once, so
the name can now be used by the previous CanvasAppSpec type.
This also renames defaultAppSpec to defaultApp.